Better semantics for KParseResult.map
This implementation unconditionally evaluates all items in the
given list, regardless of whether any of them fail. It also uses
constant space as opposed to the O(n) space usage of the previous
implementation. The change in semantics means that partial results
are not discarded and the parser gives many more parse errors up front.