Implement initial export command
This adds a complete pretty printer for imperative documents, and
contains extremely incomplete support for reconstructing imports.

A bug was fixed regarding the imperative document builder: If an
element such as a paragraph was imported whilst a subsection was in the
process of being built, the subsection would be completed and then the
imported element would be added as content to the containing section
(violating the invariant that sections may only contain subsections
xor subsection content).