Reorganize and rename types, implement path enforcement
This implements path enforcement for non-XML parsers. Parsers will
refuse to import or include files that are not descendants of the
base directory. The code was reorganized and most of KSParsers was
moved into the xom and parser packages (and tests were added for the
imperative driver). The KSParserType was renamed to KSParserDriverType
to better distinguish it from the existing parser types.

Affects #33