Escape characters for LaTeX
All unacceptable characters are now replaced with U+FFFD, and
other somewhat-less-unacceptable characters are replaced with
their long forms (such as \textbackslash{} instead of plain
backslashes and so on).

It seems that only XeTeX is really capable of processing unicode
documents these days (plain LaTeX could not be persuaded to accept
U+FFFD even with \DeclareUnicodeCharacter). The use of the inputenc
package has been removed and a currently undocumented assumption has
been introduced that the user will be using XeTeX.

Affects #32